requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
flask-restx==1.0.6
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
pyflakes==3.0.1
pyrsistent==0.19.3
pyserial==3.5
pytz==2022.7.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.13.0

Flask~=2.2.2